requirements.txt
pyyaml>=6.0.1
pathlib>=1.0.1
codex-cli>=0.1.0
boto3>=1.34.0
pyspark>=3.5.0
neo4j>=5.19.0
pandas>=2.2.0
sqlalchemy>=2.0.29
psycopg2-binary>=2.9.9
great_expectations>=0.18.4
dash>=2.15.0
dash-cytoscape>=0.3.0
plotly>=5.18.0
pytest>=8.1.1
apache-airflow==2.8.1 